% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{beta.hat}
\alias{beta.hat}
\title{Fit ordinary least squares}
\usage{
beta.hat(data, indices)
}
\arguments{
\item{data}{should be a list with data$x as design matrix and data$y as response}

\item{indices}{indices of data$x dat are included in the model}
}
\description{
This function estimates the coefficient vector beta.
}
\keyword{OLS}
