% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{EBIC}
\alias{EBIC}
\title{EBIC}
\usage{
EBIC(data, indices, const)
}
\arguments{
\item{data}{should be a list with data$x as design matrix and data$y as response}

\item{indices}{indices of data$x dat are included in the model}

\item{const}{parameter for the EBIC, for BIC specify const = 0.}
}
\description{
This function computes extended BIC (EBIC).
}
\examples{
EBIC()
}
\keyword{EBIC}
