% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{EBIC_beta}
\alias{EBIC_beta}
\title{Extended BIC for given estimate beta}
\usage{
EBIC_beta(data, beta, const)
}
\arguments{
\item{data}{should be a list with data$x as design matrix and data$y as response}

\item{beta}{given estimate for coefficients of length p + 1}

\item{const}{parameter for the EBIC, for BIC specify const = 0.}
}
\description{
This function computes extended BIC for given estimate beta of length p+1
}
\examples{
EBIC_beta()
}
\keyword{EBIC}
