% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{method_res_scenario}
\alias{method_res_scenario}
\title{Evaluate simulation results (internal)}
\usage{
method_res_scenario(
  model,
  beta1,
  s0,
  data.test,
  n_test,
  time,
  beta = NULL,
  data
)
}
\arguments{
\item{model}{input}

\item{beta1}{input}

\item{s0}{input}

\item{data.test}{input}

\item{n_test}{input}

\item{time}{input}

\item{beta}{input}

\item{data}{input}
}
\description{
This helper function evaluates simulation results.
}
\keyword{simulations}
