% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simdata.toeplitz.corr}
\alias{simdata.toeplitz.corr}
\title{Simulate toeplitz correlated data}
\usage{
simdata.toeplitz.corr(n, p, beta, sigma.normal, corr = 0)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}

\item{beta}{true coefficient vector (of size p)}

\item{sigma.normal}{epsilon}

\item{corr}{base correlation}
}
\description{
This function simulates toeplitz correlated data.
}
\examples{
simdata.toeplitz.corr()
}
\keyword{toeplitz}
