% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubBoost.R
\name{SubBoost}
\alias{SubBoost}
\title{SubBoost}
\usage{
SubBoost(
  data,
  Iter,
  size.fixed = NULL,
  tau = 0.01,
  const = 0,
  savings = 1,
  family = "normal",
  s_max = 20,
  automatic.stopping = TRUE,
  plotting = FALSE
)
}
\arguments{
\item{data}{should be a list with data$x as design matrix and data$y as response}

\item{Iter}{iterations}

\item{size.fixed}{default is set to NULL}

\item{tau}{parameter tau - default is set to 0.01}

\item{const}{parameter const - default is set to 0}

\item{savings}{default is set to 1}

\item{family}{default is set to "normal"}

\item{s_max}{default is set to 20}

\item{automatic.stopping}{default is set to TRUE}

\item{plotting}{default is set to FALSE}
}
\description{
This function implements the subspace boosting algorithm (SubBoost).
}
\keyword{SubBoost}
\examples{
require(TH.data)

# input data format:
# list with design matrix in data$x and response vector in data$y
data <- list()
data$x <- as.matrix(bodyfat[,-2])
data$y <- as.vector(bodyfat$DEXfat)

# constant (gamma) in EBIC
const <- 0 # classical BIC

# learning rate
tau <- 0.01

# (maximum) number of iterations
Iter <- 1000

# SubBoost (only applicable for low-dimensional settings, e.g. p<=20)
outputSub <- SubBoost(data = data, Iter = Iter, const = const, tau = tau)
outputSub$selected # selected variables by SubBoost
outputSub$coef # estimated coefficient vector by SubBoost
}

