% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{y.hat}
\alias{y.hat}
\title{Get y hat values}
\usage{
y.hat(data, beta.hat)
}
\arguments{
\item{data}{should be a list with data$x as design matrix}

\item{beta.hat}{coefficient vector}
}
\description{
This function estimates the y given x and coefficient vector beta.
}
\keyword{OLS}
