% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaSubBoost.R
\name{AdaSubBoost}
\alias{AdaSubBoost}
\title{AdaSubBoost and RSubBoost}
\usage{
AdaSubBoost(
  data,
  Iter,
  K = 100,
  q = 10,
  size.fixed = NULL,
  tau = 0.01,
  const = 0,
  savings = 1,
  U_C = 20,
  family = "normal",
  conservative = TRUE,
  update = "S",
  adaptive = TRUE,
  s_max = 20,
  nstop = Iter,
  automatic.stopping = TRUE,
  marginal.screening = FALSE,
  plotting = FALSE
)
}
\arguments{
\item{data}{should be a list with data$x as design matrix and data$y as response}

\item{Iter}{iterations}

\item{K}{parameter K - default is set to 100}

\item{q}{parameter q - default is set to 10}

\item{size.fixed}{default is set to NULL}

\item{tau}{parameter tau - default is set to 0.01}

\item{const}{parameter const - default is set to 0}

\item{savings}{default is set to 1}

\item{U_C}{parameter is set to 20}

\item{family}{default is set to "normal"}

\item{conservative}{default is set to TRUE}

\item{update}{default is set to "S"}

\item{adaptive}{default is set to TRUE (AdaSubBoost), for RSubBoost specify as FALSE}

\item{s_max}{default is set to 20}

\item{nstop}{default is set to  Iter}

\item{marginal.screening}{default is set to FALSE}

\item{plotting}{default is set to FALSE}

\item{automated.stopping}{default is set to TRUE}
}
\description{
This function implements the adaptive subspace boosting algorithm (AdaSubBoost) as well as random subspace boosting (RSubBoost).
}
\examples{
AdaSubBoost()
}
\keyword{AdaSubBoost,}
\keyword{RSubBoost}
