% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubBoost.R
\name{SubBoost}
\alias{SubBoost}
\title{SubBoost}
\usage{
SubBoost(
  data,
  Iter,
  size.fixed = NULL,
  tau = 0.01,
  const = 0,
  savings = 1,
  family = "normal",
  s_max = 20,
  automatic.stopping = TRUE,
  plotting = FALSE
)
}
\arguments{
\item{data}{should be a list with data$x as design matrix and data$y as response}

\item{Iter}{iterations}

\item{size.fixed}{default is set to NULL}

\item{tau}{parameter tau - default is set to 0.01}

\item{const}{parameter const - default is set to 0}

\item{savings}{default is set to 1}

\item{family}{default is set to "normal"}

\item{s_max}{default is set to 20}

\item{automatic.stopping}{default is set to TRUE}

\item{plotting}{default is set to FALSE}
}
\description{
This function implements the subspace boosting algorithm (SubBoost).
}
\examples{
SubBoost()
}
\keyword{SubBoost}
